Feature/432 confirm dialog

+++ web/containers/ClassConnectionsManager/diagram.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useContext } from 'react';
 import { IClassConnection, StyledDialogBody } from './index';
 import size from 'lodash/size';
 import find from 'lodash/find';
@@ -23,6 +23,7 @@ import withGlobalOptionsConsumer from '../../hocomponents/withGlobalOptionsConsu
 import useMount from 'react-use/lib/useMount';
 import BooleanField from '../../components/Field/boolean';
 import withMethodsConsumer from '../../hocomponents/withMethodsConsumer';
+import { InitialContext } from '../../context/init';
 
 export interface IClassConnectionsDiagramProps {
     connection: IClassConnection[];
@@ -225,6 +226,7 @@ const ClassConnectionsDiagram: React.FC<IClassConnectionsDiagramProps> = ({
     const [manageDialog, setManageDialog] = useState<IManageDialog>({});
     const [hasLast, setHasLast] = useState<boolean>(false);
     const [mapperDialog, setMapperDialog] = useState({});
+    const initContext = useContext(InitialContext);
 
     const isConnectorValid = () => {
         return manageDialog.isMapper
@@ -359,10 +361,12 @@ const ClassConnectionsDiagram: React.FC<IClassConnectionsDiagramProps> = ({
                                                         icon="trash"
                                                         intent="danger"
                                                         onClick={() => {
-                                                            setManageDialog(current => ({
-                                                                ...current,
-                                                                mapper: null,
-                                                            }));
+                                                            initContext.confirmAction('ConfirmRemoveMapper', () =>
+                                                                setManageDialog(current => ({
+                                                                    ...current,
+                                                                    mapper: null,
+                                                                }))
+                                                            );
                                                         }}
                                                     />
                                                 </>
@@ -481,10 +485,12 @@ const ClassConnectionsDiagram: React.FC<IClassConnectionsDiagramProps> = ({
                                                             icon="trash"
                                                             intent="danger"
                                                             onClick={() => {
-                                                                setManageDialog(current => ({
-                                                                    ...current,
-                                                                    trigger: null,
-                                                                }));
+                                                                initContext.confirmAction('ConfirmRemoveTrigger', () =>
+                                                                    setManageDialog(current => ({
+                                                                        ...current,
+                                                                        trigger: null,
+                                                                    }))
+                                                                );
                                                             }}
                                                         />
                                                     )}
@@ -586,15 +592,17 @@ const ClassConnectionsDiagram: React.FC<IClassConnectionsDiagramProps> = ({
                                                 <Button
                                                     small
                                                     minimal
-                                                    icon={<Icon icon={'trash'} iconSize={12} />}
+                                                    icon={<Icon icon={'trash'} intent="danger" iconSize={12} />}
                                                     onClick={() => {
-                                                        onAddConnector(
-                                                            connectionName,
-                                                            {
-                                                                index,
-                                                                isEditing: true,
-                                                            },
-                                                            true
+                                                        initContext.confirmAction('ConfirmRemoveMapper', () =>
+                                                            onAddConnector(
+                                                                connectionName,
+                                                                {
+                                                                    index,
+                                                                    isEditing: true,
+                                                                },
+                                                                true
+                                                            )
                                                         );
                                                     }}
                                                 />
@@ -656,12 +664,14 @@ const ClassConnectionsDiagram: React.FC<IClassConnectionsDiagramProps> = ({
                                     <Tooltip content={t('RemoveConnector')}>
                                         <Button
                                             onClick={() => {
-                                                onDeleteConnector(connectionName, index);
-                                                // If this was the last connector
-                                                if (conn.isLast) {
-                                                    // Remove the last flag
-                                                    setHasLast(false);
-                                                }
+                                                initContext.confirmAction('ConfirmRemoveConnector', () => {
+                                                    onDeleteConnector(connectionName, index);
+                                                    // If this was the last connector
+                                                    if (conn.isLast) {
+                                                        // Remove the last flag
+                                                        setHasLast(false);
+                                                    }
+                                                });
                                             }}
                                             minimal
                                             icon="trash"
